pyproject.toml
[project]
name = "pyfulmen"
version = "0.1.0"
description = "Python Fulmen libraries"
authors = [
    {name = "3 Leaps Team", email = "dev@3leaps.net"}
]
requires-python = ">=3.12"
license = {text = "MIT"}
readme = "README.md"
keywords = ["fulmen", "library", "enterprise"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    "jsonschema>=4.25.1",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyfulmen"]

[tool.hatch.build.targets.wheel.force-include]
"src/pyfulmen/py.typed" = "pyfulmen/py.typed"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
